safari-600.7-branch/LayoutTests/media/media-source/media-source-overlapping-append-buffered-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(samples))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(samples))
EVENT(updateend)
EXPECTED (video.buffered.length == '1') OK
EXPECTED (video.buffered.start(0).toFixed(3) == '1') OK
EXPECTED (video.buffered.end(0).toFixed(3) == '7') OK
END OF TEST
